[PATCH] s2io.h: gcc-3.5 build fix
In file included from drivers/net/s2io.c:68:
drivers/net/s2io.h: In function `readq':
drivers/net/s2io.h:757: error: invalid lvalue in assignment
drivers/net/s2io.h:758: error: invalid lvalue in assignment

+++ drivers/net/s2io.h
@@ -754,8 +754,8 @@ static inline u64 readq(void *addr)
 {
 	u64 ret = 0;
 	ret = readl(addr + 4);
-	(u64) ret <<= 32;
-	(u64) ret |= readl(addr);
+	ret <<= 32;
+	ret |= readl(addr);
 
 	return ret;
 }
